Backend/file-format-identification-service/requirements.txt
Flask==2.1.0
python-magic==0.4.24